# dist/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Hyperledger Explorer REST API Swagger",
        "description": "Rest API for fabric",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Hyperledger Team"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080"
        },
        {
            "url": "https://localhost:8080"
        }
    ],
...
    "paths": {
...
        "/api/blockAndTxList/{channel_genesis_hash}/{number}": {
            "get": {
                "tags": ["Transactions"],
                "summary": "Query to fetch blocks and transaction list",
                "operationId": "getBlockAndTxList",
                "parameters": [
                    {
                        "name": "channel_genesis_hash",
                        "in": "path",
                        "description": "Please give the Hash Value",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "number",
                        "in": "path",
                        "description": "Please give the number value",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {}
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {}
                    },
                    "401": {
                        "description": "UnAuthorized",
                        "content": {}
                    },
                    "404": {
                        "description": "Not found",
                        "content": {}
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}